implementation optional removal of messages pop-up
I've been having this problem that every time Twitter loaded, the somewhat recent addition of the messages to the bottom right corner of the screen would expand automatically. I suspect this is one of those mobile features they gave up on implementing properly half way through. 
"Since I've already got an extension right here, why not make it remove the messages thingy" is what I thought and set out to do.
And now it's an optional feature, found in the options tab of this add-on.
I can't tell you the default value, because I don't know so check it :)